requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
Pillow==10.2.0
APScheduler==3.10.4
python-dotenv==1.0.0